Update README.md
My system Python is version 3.9.4. This led to me having issues installing several of the libraries in requirements.txt. This happened because the mentioned versions of several libraries only support Python up-to version 3.8.x (mine was 3.9+).
Naturally, virtualenv is the best solution to this.
Since most of the Python users use the latest version, they should be up updated about this.